backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
asyncpg==0.30.0
certifi==2025.1.31
click==8.1.8
exceptiongroup==1.2.2
Faker==36.1.1
fastapi==0.115.10
greenlet==3.1.1
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
orjson==3.10.15
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
rfc3986==1.5.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.23
starlette==0.46.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.15.0
